Migrated scheduler into a component
