+++ flexget/api/__init__.py
@@ -3,5 +3,5 @@ from __future__ import unicode_literals, division, absolute_import  # noqa
 from .app import api_app, api, APIResource, APIClient  # noqa
 from .core import authentication, cached, database, plugins, server, tasks, user, format_checker  # noqa
 from .plugins import (
-    tvdb_lookup, tvmaze_lookup, tmdb_lookup, rejected, schedule, variables  # noqa
+    tvdb_lookup, tvmaze_lookup, tmdb_lookup, variables  # noqa
 )
