+++ flexget/components/rejected/api.py
@@ -8,8 +8,14 @@ from math import ceil
 from sqlalchemy.orm.exc import NoResultFound
 
 from flexget.api import api, APIResource
-from flexget.api.app import base_message_schema, success_response, etag, NotFoundError, pagination_headers
-from flexget.plugins.filter.remember_rejected import RememberEntry, get_rejected
+from flexget.api.app import (
+    base_message_schema,
+    success_response,
+    etag,
+    NotFoundError,
+    pagination_headers,
+)
+from . import db
 
 log = logging.getLogger('rejected')
 
@@ -24,7 +30,7 @@ def rejected_entry_to_dict(entry):
         'title': entry.title,
         'url': entry.url,
         'rejected_by': entry.rejected_by,
-        'reason': entry.reason
+        'reason': entry.reason,
     }
 
 
@@ -38,17 +44,20 @@ class ObjectsContainer(object):
             'added': {'type': 'string', 'format': 'date-time'},
             'reason': {'type': 'string'},
             'expires': {'type': 'string', 'format': 'date-time'},
-            'rejected_by': {'type': 'string'}
+            'rejected_by': {'type': 'string'},
         },
         'required': ['id', 'title', 'url', 'added', 'reason', 'expires', 'rejected_by'],
-        'additionalProperties': False
-
+        'additionalProperties': False,
     }
     rejected_entries_list_object = {'type': 'array', 'items': rejected_entry_object}
 
 
-rejected_entry_schema = api.schema_model('rejected_failed_entry_schema', ObjectsContainer.rejected_entry_object)
-rejected_entries_list_schema = api.schema_model('rejected_entries_list_schema', ObjectsContainer.rejected_entries_list_object)
+rejected_entry_schema = api.schema_model(
+    'rejected_failed_entry_schema', ObjectsContainer.rejected_entry_object
+)
+rejected_entries_list_schema = api.schema_model(
+    'rejected_entries_list_schema', ObjectsContainer.rejected_entries_list_object
+)
 
 sort_choices = ('added', 'id', 'title', 'url', 'expires', 'rejected_by', 'reason')
 rejected_parser = api.pagination_parser(sort_choices=sort_choices)
@@ -88,15 +97,15 @@ class Rejected(APIResource):
             'stop': stop,
             'descending': descending,
             'sort_by': sort_by,
-            'session': session
+            'session': session,
         }
 
-        total_items = get_rejected(session, count=True)
+        total_items = db.get_rejected(session, count=True)
 
         if not total_items:
             return jsonify([])
 
-        failed_entries = [rejected_entry_to_dict(reject) for reject in get_rejected(**kwargs)]
+        failed_entries = [rejected_entry_to_dict(reject) for reject in db.get_rejected(**kwargs)]
 
         total_pages = int(ceil(total_items / float(per_page)))
 
@@ -120,7 +129,7 @@ class Rejected(APIResource):
     @api.response(200, model=base_message_schema)
     def delete(self, session=None):
         """ Clears all rejected entries"""
-        entries = session.query(RememberEntry).delete()
+        entries = session.query(db.RememberEntry).delete()
         if entries:
             session.commit()
             self.manager.config_changed()
@@ -135,7 +144,11 @@ class RejectedEntry(APIResource):
     def get(self, rejected_entry_id, session=None):
         """ Returns a rejected entry """
         try:
-            entry = session.query(RememberEntry).filter(RememberEntry.id == rejected_entry_id).one()
+            entry = (
+                session.query(db.RememberEntry)
+                .filter(db.RememberEntry.id == rejected_entry_id)
+                .one()
+            )
         except NoResultFound:
             raise NotFoundError('rejected entry ID %d not found' % rejected_entry_id)
         return jsonify(rejected_entry_to_dict(entry))
@@ -144,7 +157,11 @@ class RejectedEntry(APIResource):
     def delete(self, rejected_entry_id, session=None):
         """ Deletes a rejected entry """
         try:
-            entry = session.query(RememberEntry).filter(RememberEntry.id == rejected_entry_id).one()
+            entry = (
+                session.query(db.RememberEntry)
+                .filter(db.RememberEntry.id == rejected_entry_id)
+                .one()
+            )
         except NoResultFound:
             raise NotFoundError('rejected entry ID %d not found' % rejected_entry_id)
         session.delete(entry)
