+++ flexget/tests/api_tests/test_rejected_api.py
@@ -5,9 +5,9 @@ import copy
 from datetime import datetime
 
 from flexget.api.app import base_message
-from flexget.api.plugins.rejected import ObjectsContainer as OC
+from flexget.components.rejected.api import ObjectsContainer as OC
 from flexget.manager import Session
-from flexget.plugins.filter.remember_rejected import RememberEntry, RememberTask
+from flexget.components.rejected.db import RememberEntry, RememberTask
 from flexget.utils import json
 from flexget.utils.tools import parse_timedelta
