+++ flexget/tests/api_tests/test_schedule_api.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals, division, absolute_import
 from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 from flexget.api.app import base_message
-from flexget.api.plugins.schedule import ObjectsContainer as OC
+from flexget.components.scheduler.api import ObjectsContainer as OC
 from flexget.manager import Manager
 from flexget.utils import json
 from mock import patch
